refactor(jobs): modular architecture with JobBackend trait
Split the monolithic 4,483-line src/stdlib/jobs.rs into a clean modular
architecture:

  src/stdlib/jobs/
    mod.rs          — Public API, types, statics, dispatch layer
    backend.rs      — JobBackend trait definition + ClaimedJob type
    memory.rs       — In-memory backend (queue state, enqueue, claim, etc.)
    postgres.rs     — PostgreSQL backend (pool, migrations, LISTEN/NOTIFY)
    redis_backend.rs — Redis Streams backend (consumer groups, XCLAIM)
    worker.rs       — Worker loop implementations for all backends

Key improvements:
- Each backend is self-contained in its own file
- JobBackend trait defines the pluggable interface for future unification
- Worker loops organized in one file with shared execution logic
- mod.rs is a thin dispatch layer (~800 lines vs 4,483)
- No behavior changes: all 1,063 tests pass, all docs validate

Zero breaking changes to public API or user-facing syntax.
